// autogpt_platform/frontend/src/app/api/__generated__/models/storeWaitlistEntry.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreWaitlistEntryStoreListing } from "./storeWaitlistEntryStoreListing";
import type { StoreWaitlistEntryOwner } from "./storeWaitlistEntryOwner";
import type { StoreWaitlistEntryVideoUrl } from "./storeWaitlistEntryVideoUrl";
import type { StoreWaitlistEntryAgentOutputDemoUrl } from "./storeWaitlistEntryAgentOutputDemoUrl";

export interface StoreWaitlistEntry {
  waitlist_id: string;
  storeListing?: StoreWaitlistEntryStoreListing;
  owner?: StoreWaitlistEntryOwner;
  slug: string;
  name: string;
  subHeading: string;
  videoUrl?: StoreWaitlistEntryVideoUrl;
  agentOutputDemoUrl?: StoreWaitlistEntryAgentOutputDemoUrl;
  imageUrls: string[];
  description: string;
  categories: string[];
}